Rename LoginServlet to LoginCheckServlet
